refactor(main-page)/ main-page에 사용되는 스크립트를 기능별로 분리하였습니다.
